requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
pycparser==2.20
python-telegram-bot==12.7
six==1.14.0
tornado==6.0.4
